Refactor daemon test helpers into a support module
