+++ crates/daemon/src/tests.rs
@@ -5,129 +5,21 @@ use std::borrow::Cow;
 use std::ffi::{OsStr, OsString};
 use std::fs;
 use std::io::{BufRead, BufReader, Read, Write};
-use std::net::{IpAddr, Ipv4Addr, Ipv6Addr, SocketAddr, TcpListener, TcpStream};
+use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};
 use std::num::{NonZeroU32, NonZeroU64};
 use std::path::{Path, PathBuf};
-use std::sync::Mutex;
 use std::sync::{Arc, Barrier};
 use std::thread;
 use std::time::Duration;
 use tempfile::{NamedTempFile, tempdir};
 
-#[cfg(unix)]
-use std::os::unix::fs::PermissionsExt;
-
-const RSYNCD: &str = branding::daemon_program_name();
-const OC_RSYNC_D: &str = branding::oc_daemon_program_name();
-
-fn base_module(name: &str) -> ModuleDefinition {
-    ModuleDefinition {
-        name: String::from(name),
-        path: PathBuf::from("/srv/module"),
-        comment: None,
-        hosts_allow: Vec::new(),
-        hosts_deny: Vec::new(),
-        auth_users: Vec::new(),
-        secrets_file: None,
-        bandwidth_limit: None,
-        bandwidth_limit_specified: false,
-        bandwidth_burst: None,
-        bandwidth_burst_specified: false,
-        bandwidth_limit_configured: false,
-        refuse_options: Vec::new(),
-        read_only: true,
-        numeric_ids: false,
-        uid: None,
-        gid: None,
-        timeout: None,
-        listable: true,
-        use_chroot: true,
-        max_connections: None,
-    }
-}
-
-static ENV_LOCK: Mutex<()> = Mutex::new(());
-
-fn with_test_secrets_candidates<F, R>(candidates: Vec<PathBuf>, func: F) -> R
-where
-    F: FnOnce() -> R,
-{
-    TEST_SECRETS_CANDIDATES.with(|cell| {
-        let previous = cell.replace(Some(candidates));
-        let result = func();
-        cell.replace(previous);
-        result
-    })
-}
-
-fn with_test_secrets_env<F, R>(override_value: Option<TestSecretsEnvOverride>, func: F) -> R
-where
-    F: FnOnce() -> R,
-{
-    TEST_SECRETS_ENV.with(|cell| {
-        let previous = cell.replace(override_value);
-        let result = func();
-        cell.replace(previous);
-        result
-    })
-}
-
-fn allocate_test_port() -> u16 {
-    const START: u16 = 40_000;
-    const RANGE: u32 = 20_000;
-    const STATE_SIZE: u64 = 4;
-
-    let mut path = std::env::temp_dir();
-    path.push("rsync-daemon-test-port.lock");
-
-    let mut file = OpenOptions::new()
-        .read(true)
-        .write(true)
-        .create(true)
-        .truncate(false)
-        .open(&path)
-        .expect("open port allocator state");
-
-    file.lock_exclusive().expect("lock port allocator state");
-    file.seek(SeekFrom::Start(0))
-        .expect("rewind port allocator state");
-
-    let mut counter_bytes = [0u8; STATE_SIZE as usize];
-    let read = file
-        .read(&mut counter_bytes)
-        .expect("read port allocator state");
-    let mut counter = if read == counter_bytes.len() {
-        u32::from_le_bytes(counter_bytes)
-    } else {
-        0
-    };
-
-    for _ in 0..RANGE {
-        let offset = (counter % RANGE) as u16;
-        counter = counter.wrapping_add(1);
-
-        file.seek(SeekFrom::Start(0))
-            .expect("rewind port allocator state");
-        file.write_all(&counter.to_le_bytes())
-            .expect("persist port allocator state");
-        file.set_len(STATE_SIZE)
-            .expect("truncate port allocator state");
-        file.flush().expect("flush port allocator state");
-
-        let candidate = START + offset;
-        if let Ok(listener) = TcpListener::bind((Ipv4Addr::LOCALHOST, candidate)) {
-            drop(listener);
-            return candidate;
-        }
-    }
-
-    panic!("failed to allocate a free test port");
-}
+mod support;
+use support::*;
 
 #[test]
 fn parse_auth_user_list_trims_and_deduplicates_case_insensitively() {
-    let users = parse_auth_user_list(" alice,BOB, alice ,  Carol ")
-        .expect("parse non-empty user list");
+    let users =
+        parse_auth_user_list(" alice,BOB, alice ,  Carol ").expect("parse non-empty user list");
     assert_eq!(users, ["alice", "BOB", "Carol"]);
 
     let err = parse_auth_user_list(" , ,  ").expect_err("blank list rejected");
@@ -182,18 +74,10 @@ fn parse_max_connections_directive_handles_zero_and_positive() {
     assert_eq!(parse_max_connections_directive("0"), Some(None));
 
     let expected = NonZeroU32::new(25).expect("non-zero");
-    assert_eq!(
-        parse_max_connections_directive("25"),
-        Some(Some(expected))
-    );
+    assert_eq!(parse_max_connections_directive("25"), Some(Some(expected)));
 
     assert_eq!(parse_max_connections_directive("-1"), None);
     assert_eq!(parse_max_connections_directive("invalid"), None);
-}
-
-struct EnvGuard {
-    key: &'static str,
-    previous: Option<OsString>,
 }
 
 #[test]
@@ -234,40 +118,6 @@ fn connection_status_messages_describe_active_sessions() {
     assert_eq!(format_connection_status(0), "Idle; waiting for connections");
     assert_eq!(format_connection_status(1), "Serving 1 connection");
     assert_eq!(format_connection_status(3), "Serving 3 connections");
-}
-
-#[allow(unsafe_code)]
-impl EnvGuard {
-    fn set(key: &'static str, value: &OsStr) -> Self {
-        let previous = std::env::var_os(key);
-        unsafe {
-            std::env::set_var(key, value);
-        }
-        Self { key, previous }
-    }
-
-    fn remove(key: &'static str) -> Self {
-        let previous = std::env::var_os(key);
-        unsafe {
-            std::env::remove_var(key);
-        }
-        Self { key, previous }
-    }
-}
-
-#[allow(unsafe_code)]
-impl Drop for EnvGuard {
-    fn drop(&mut self) {
-        if let Some(value) = self.previous.take() {
-            unsafe {
-                std::env::set_var(self.key, value);
-            }
-        } else {
-            unsafe {
-                std::env::remove_var(self.key);
-            }
-        }
-    }
 }
 
 #[test]
@@ -403,16 +253,6 @@ fn configured_fallback_binary_supports_auto_value() {
     assert_eq!(configured_fallback_binary(), Some(OsString::from("rsync")));
 }
 
-#[cfg(unix)]
-fn write_executable_script(path: &Path, contents: &str) {
-    std::fs::write(path, contents).expect("write script");
-    let mut permissions = std::fs::metadata(path)
-        .expect("script metadata")
-        .permissions();
-    permissions.set_mode(0o755);
-    std::fs::set_permissions(path, permissions).expect("set script permissions");
-}
-
 #[test]
 fn advertised_capability_lines_empty_without_modules() {
     assert!(advertised_capability_lines(&[]).is_empty());
@@ -439,49 +279,6 @@ fn advertised_capability_lines_include_authlist_when_required() {
         advertised_capability_lines(&[module]),
         vec![String::from("modules authlist")]
     );
-}
-
-fn module_with_host_patterns(allow: &[&str], deny: &[&str]) -> ModuleDefinition {
-    ModuleDefinition {
-        name: String::from("module"),
-        path: PathBuf::from("/srv/module"),
-        comment: None,
-        hosts_allow: allow
-            .iter()
-            .map(|pattern| HostPattern::parse(pattern).expect("parse allow pattern"))
-            .collect(),
-        hosts_deny: deny
-            .iter()
-            .map(|pattern| HostPattern::parse(pattern).expect("parse deny pattern"))
-            .collect(),
-        auth_users: Vec::new(),
-        secrets_file: None,
-        bandwidth_limit: None,
-        bandwidth_limit_specified: false,
-        bandwidth_burst: None,
-        bandwidth_burst_specified: false,
-        bandwidth_limit_configured: false,
-        refuse_options: Vec::new(),
-        read_only: true,
-        numeric_ids: false,
-        uid: None,
-        gid: None,
-        timeout: None,
-        listable: true,
-        use_chroot: true,
-        max_connections: None,
-    }
-}
-
-fn run_with_args<I, S>(args: I) -> (i32, Vec<u8>, Vec<u8>)
-where
-    I: IntoIterator<Item = S>,
-    S: Into<OsString>,
-{
-    let mut stdout = Vec::new();
-    let mut stderr = Vec::new();
-    let code = run(args, &mut stdout, &mut stderr);
-    (code, stdout, stderr)
 }
 
 #[test]
@@ -4584,28 +4381,4 @@ fn clap_parse_error_is_reported_via_message() {
 
     let rendered = String::from_utf8(stderr).expect("diagnostic is valid UTF-8");
     assert!(rendered.contains(error.to_string().trim()));
-}
-
-fn connect_with_retries(port: u16) -> TcpStream {
-    const INITIAL_BACKOFF: Duration = Duration::from_millis(20);
-    const MAX_BACKOFF: Duration = Duration::from_millis(200);
-    const TIMEOUT: Duration = Duration::from_secs(15);
-
-    let target = SocketAddr::from((Ipv4Addr::LOCALHOST, port));
-    let deadline = Instant::now() + TIMEOUT;
-    let mut backoff = INITIAL_BACKOFF;
-
-    loop {
-        match TcpStream::connect_timeout(&target, backoff) {
-            Ok(stream) => return stream,
-            Err(error) => {
-                if Instant::now() >= deadline {
-                    panic!("failed to connect to daemon within timeout: {error}");
-                }
-
-                thread::sleep(backoff);
-                backoff = (backoff.saturating_mul(2)).min(MAX_BACKOFF);
-            }
-        }
-    }
 }
